+++ src/Analysis/Engine/Impl/Interpreter/Ast/AstPythonInterpreter.cs
@@ -24,27 +24,27 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.PythonTools.Analysis;
 using Microsoft.PythonTools.Analysis.Infrastructure;
+using Microsoft.PythonTools.Parsing;
 
 namespace Microsoft.PythonTools.Interpreter.Ast {
     internal class AstPythonInterpreter : IPythonInterpreter2, IModuleContext, ICanFindModuleMembers {
         private readonly AstPythonInterpreterFactory _factory;
         private readonly Dictionary<BuiltinTypeId, IPythonType> _builtinTypes;
-        private PythonAnalyzer _analyzer;
-        private AstScrapedPythonModule _builtinModule;
-        private IReadOnlyList<string> _builtinModuleNames;
+        private readonly string _workspaceRoot;
         private readonly ConcurrentDictionary<string, IPythonModule> _modules;
         private readonly AstPythonBuiltinType _noneType;
 
-        internal readonly AnalysisLogWriter _log;
+        private PythonAnalyzer _analyzer;
+        private AstScrapedPythonModule _builtinModule;
+        private IReadOnlyList<string> _builtinModuleNames;
 
-        private readonly object _userSearchPathsLock = new object();
-        private IReadOnlyList<string> _userSearchPaths;
         private IReadOnlyDictionary<string, string> _userSearchPathPackages;
-        private HashSet<string> _userSearchPathImported;
+        private ConcurrentBag<string> _userSearchPathImported = new ConcurrentBag<string>();
 
-        public AstPythonInterpreter(AstPythonInterpreterFactory factory, AnalysisLogWriter log = null) {
+        public AstPythonInterpreter(AstPythonInterpreterFactory factory, string workspaceRoot, AnalysisLogWriter log = null) {
             _factory = factory ?? throw new ArgumentNullException(nameof(factory));
-            _log = log;
+            _workspaceRoot = workspaceRoot;
+            Log = log;
             _factory.ImportableModulesChanged += Factory_ImportableModulesChanged;
             _modules = new ConcurrentDictionary<string, IPythonModule>();
             _builtinTypes = new Dictionary<BuiltinTypeId, IPythonType>();
@@ -57,6 +57,13 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
             _factory.ImportableModulesChanged -= Factory_ImportableModulesChanged;
         }
 
+
+        public event EventHandler ModuleNamesChanged;
+        public IModuleContext CreateModuleContext() => this;
+        public IPythonInterpreterFactory Factory => _factory;
+        public string BuiltinModuleName => BuiltinTypeId.Unknown.GetModuleName(_factory.Configuration.Version.ToLanguageVersion());
+        public AnalysisLogWriter Log { get; }
+
         private void Factory_ImportableModulesChanged(object sender, EventArgs e) {
             _modules.Clear();
             ModuleNamesChanged?.Invoke(this, EventArgs.Empty);
@@ -65,12 +72,6 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
         public void AddUnimportableModule(string moduleName) {
             _modules[moduleName] = new SentinelModule(moduleName, false);
         }
-
-        public event EventHandler ModuleNamesChanged;
-
-        public IModuleContext CreateModuleContext() => this;
-        public IPythonInterpreterFactory Factory => _factory;
-        public string BuiltinModuleName => _factory.BuiltinModuleName;
 
         public IPythonType GetBuiltinType(BuiltinTypeId id) {
             if (id < 0 || id > BuiltinTypeIdExtensions.LastTypeId) {
@@ -100,58 +101,32 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
         }
 
         private async Task<IReadOnlyDictionary<string, string>> GetUserSearchPathPackagesAsync(CancellationToken cancellationToken) {
-            _log?.Log(TraceLevel.Verbose, "GetUserSearchPathPackagesAsync");
-            var ussp = _userSearchPathPackages;
-            if (ussp == null) {
-                IReadOnlyList<string> usp;
-                lock (_userSearchPathsLock) {
-                    usp = _userSearchPaths;
-                    ussp = _userSearchPathPackages;
-                }
-                if (ussp != null || usp == null || !usp.Any()) {
-                    return ussp;
-                }
-
-                _log?.Log(TraceLevel.Verbose, "GetImportableModulesAsync");
-                var requireInitPy = ModulePath.PythonVersionRequiresInitPyFiles(_factory.Configuration.Version);
-                ussp = await AstPythonInterpreterFactory.GetImportableModulesAsync(usp, requireInitPy, cancellationToken);
-                lock (_userSearchPathsLock) {
-                    if (_userSearchPathPackages == null) {
-                        _userSearchPathPackages = ussp;
-                    } else {
-                        ussp = _userSearchPathPackages;
-                    }
-                }
+            Log?.Log(TraceLevel.Verbose, "GetUserSearchPathPackagesAsync");
+            if (_userSearchPathPackages == null) {
+                _userSearchPathPackages = await _factory.ModuleResolution.GetPackagesFromSearchPathsAsync(_analyzer.GetSearchPaths(), cancellationToken);
+                Log?.Log(TraceLevel.Verbose, "GetPackagesFromSearchPathsAsync", _userSearchPathPackages.Keys.Cast<object>().ToArray());
             }
-
-            _log?.Log(TraceLevel.Verbose, "GetUserSearchPathPackagesAsync", ussp.Keys.Cast<object>().ToArray());
-            return ussp;
+            return _userSearchPathPackages;
         }
 
-        private void ImportedFromUserSearchPath(string name) {
-            lock (_userSearchPathsLock) {
-                if (_userSearchPathImported == null) {
-                    _userSearchPathImported = new HashSet<string>();
+        private async Task<ModulePath?> FindModuleOnDiskAsync(string name, CancellationToken cancellationToken) {
+            IReadOnlyList<string> searchPaths = new[] { _workspaceRoot };
+
+            var packages = await _factory.ModuleResolution.GetPackagesFromSearchPathsAsync(searchPaths, cancellationToken);
+            if (packages == null) {
+                searchPaths = _analyzer.GetSearchPaths();
+                if (searchPaths == null || searchPaths.Count == 0) {
+                    return null;
                 }
-                _userSearchPathImported.Add(name);
+                packages = await _factory.ModuleResolution.GetPackagesFromSearchPathsAsync(searchPaths, cancellationToken);
             }
-        }
-
-        private async Task<ModulePath?> FindModuleInUserSearchPathAsync(string name, CancellationToken cancel) {
-            var searchPaths = _userSearchPaths;
-
-            if (searchPaths == null || searchPaths.Count == 0) {
-                return null;
-            }
-
-            var packages = await GetUserSearchPathPackagesAsync(cancel);
 
             var i = name.IndexOf('.');
             var firstBit = i < 0 ? name : name.Remove(i);
             string searchPath;
 
             ModulePath mp;
-            Func<string, bool> isPackage = _factory.IsPackage;
+            Func<string, bool> isPackage = _factory.ModuleResolution.IsPackage;
             if (firstBit.EndsWithOrdinal("-stubs", ignoreCase: true)) {
                 isPackage = Directory.Exists;
             }
@@ -159,14 +134,14 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
             var requireInitPy = ModulePath.PythonVersionRequiresInitPyFiles(_factory.Configuration.Version);
             if (packages != null && packages.TryGetValue(firstBit, out searchPath) && !string.IsNullOrEmpty(searchPath)) {
                 if (ModulePath.FromBasePathAndName_NoThrow(searchPath, name, isPackage, null, requireInitPy, out mp, out _, out _, out _)) {
-                    ImportedFromUserSearchPath(name);
+                    _userSearchPathImported.Add(name);
                     return mp;
                 }
             }
 
             foreach (var sp in searchPaths.MaybeEnumerate()) {
                 if (ModulePath.FromBasePathAndName_NoThrow(sp, name, isPackage, null, requireInitPy, out mp, out _, out _, out _)) {
-                    ImportedFromUserSearchPath(name);
+                    _userSearchPathImported.Add(name);
                     return mp;
                 }
             }
@@ -176,7 +151,7 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
 
         public IList<string> GetModuleNames() {
             var ussp = GetUserSearchPathPackagesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
-            var ssp = _factory.GetImportableModulesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
+            var ssp = _factory.ModuleResolution.GetImportableModulesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
             var bmn = _builtinModuleNames;
             if (bmn == null && _builtinModule != null) {
                 var builtinModules = (_builtinModule as IBuiltinPythonModule)?.GetAnyMember("__builtin_module_names__");
@@ -208,11 +183,11 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
 
             Debug.Assert(_analyzer != null);
 
-            var ctxt = new AstPythonInterpreterFactory.TryImportModuleContext {
+            var ctxt = new TryImportModuleContext {
                 Interpreter = this,
                 ModuleCache = _modules,
                 BuiltinModule = _builtinModule,
-                FindModuleInUserSearchPathAsync = FindModuleInUserSearchPathAsync,
+                FindModuleAsync = FindModuleOnDiskAsync,
                 TypeStubPaths = _analyzer.Limits.UseTypeStubPackages ? _analyzer.GetTypeStubPaths() : null,
                 MergeTypeStubPackages = !_analyzer.Limits.UseTypeStubPackagesExclusively
             };
@@ -224,31 +199,31 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
                 // possible, but if all else fails then we'll abort and treat it as an
                 // error.
                 // (And if we've got a debugger attached, don't time out at all.)
-                AstPythonInterpreterFactory.TryImportModuleResult result;
+                TryImportModuleResult result;
                 try {
-                    result = await _factory.TryImportModuleAsync(name, ctxt, token);
+                    result = await _factory.ModuleResolution.TryImportModuleAsync(name, ctxt, token);
                 } catch (OperationCanceledException) {
-                    _log.Log(TraceLevel.Error, "ImportTimeout", name);
+                    Log.Log(TraceLevel.Error, "ImportTimeout", name);
                     Debug.Fail("Import timeout");
                     return null;
                 }
 
                 switch (result.Status) {
-                    case AstPythonInterpreterFactory.TryImportModuleResultCode.Success:
+                    case TryImportModuleResultCode.Success:
                         return result.Module;
-                    case AstPythonInterpreterFactory.TryImportModuleResultCode.ModuleNotFound:
-                        _log?.Log(TraceLevel.Info, "ImportNotFound", name);
+                    case TryImportModuleResultCode.ModuleNotFound:
+                        Log?.Log(TraceLevel.Info, "ImportNotFound", name);
                         return null;
-                    case AstPythonInterpreterFactory.TryImportModuleResultCode.NeedRetry:
-                    case AstPythonInterpreterFactory.TryImportModuleResultCode.Timeout:
+                    case TryImportModuleResultCode.NeedRetry:
+                    case TryImportModuleResultCode.Timeout:
                         break;
-                    case AstPythonInterpreterFactory.TryImportModuleResultCode.NotSupported:
-                        _log?.Log(TraceLevel.Error, "ImportNotSupported", name);
+                    case TryImportModuleResultCode.NotSupported:
+                        Log?.Log(TraceLevel.Error, "ImportNotSupported", name);
                         return null;
                 }
             }
             // Never succeeded, so just log the error and fail
-            _log?.Log(TraceLevel.Error, "RetryImport", name);
+            Log?.Log(TraceLevel.Error, "RetryImport", name);
             return null;
         }
 
@@ -268,43 +243,33 @@ namespace Microsoft.PythonTools.Interpreter.Ast {
             return null;
         }
 
-        public void Initialize(PythonAnalyzer state) {
+        public void Initialize(PythonAnalyzer analyzer) {
             if (_analyzer != null) {
                 _analyzer.SearchPathsChanged -= Analyzer_SearchPathsChanged;
             }
 
-            _analyzer = state;
+            _analyzer = analyzer;
 
-            if (state != null) {
-                lock (_userSearchPathsLock) {
-                    _userSearchPaths = state.GetSearchPaths();
-                }
-                state.SearchPathsChanged += Analyzer_SearchPathsChanged;
-                var bm = state.BuiltinModule;
+            if (analyzer != null) {
+                analyzer.SearchPathsChanged += Analyzer_SearchPathsChanged;
+                var bm = analyzer.BuiltinModule;
                 if (!string.IsNullOrEmpty(bm?.Name)) {
-                    _modules[state.BuiltinModule.Name] = state.BuiltinModule.InterpreterModule;
+                    _modules[analyzer.BuiltinModule.Name] = analyzer.BuiltinModule.InterpreterModule;
                 }
             }
         }
 
         private void Analyzer_SearchPathsChanged(object sender, EventArgs e) {
-            lock (_userSearchPathsLock) {
-                // Remove imported modules from search paths so we will
-                // import them again.
-                foreach (var name in _userSearchPathImported.MaybeEnumerate()) {
-                    IPythonModule mod;
-                    _modules.TryRemove(name, out mod);
-                }
-                _userSearchPathImported = null;
-                _userSearchPathPackages = null;
-                _userSearchPaths = _analyzer.GetSearchPaths();
+            // Remove imported modules from search paths so we will import them again.
+            foreach (var name in _userSearchPathImported.MaybeEnumerate().ToArray()) {
+                _modules.TryRemove(name, out var mod);
             }
             ModuleNamesChanged?.Invoke(this, EventArgs.Empty);
         }
 
         public IEnumerable<string> GetModulesNamed(string name) {
             var usp = GetUserSearchPathPackagesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
-            var ssp = _factory.GetImportableModulesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
+            var ssp = _factory.ModuleResolution.GetImportableModulesAsync(CancellationToken.None).WaitAndUnwrapExceptions();
 
             var dotName = "." + name;
