+++ src/Analysis/Engine/Impl/Interpreter/Ast/TryImportModuleContext.cs
@@ -1,0 +1,33 @@
+﻿// Python Tools for Visual Studio
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABLITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
+
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.PythonTools.Analysis;
+
+namespace Microsoft.PythonTools.Interpreter.Ast {
+    public sealed class TryImportModuleContext {
+        public IPythonInterpreter Interpreter { get; set; }
+        public ConcurrentDictionary<string, IPythonModule> ModuleCache { get; set; }
+        public IPythonModule BuiltinModule { get; set; }
+        public Func<string, CancellationToken, Task<ModulePath?>> FindModuleAsync { get; set; }
+        public IReadOnlyList<string> TypeStubPaths { get; set; }
+        public bool MergeTypeStubPackages { get; set; }
+    }
+}
